LayoutTests/accessibility/ios-simulator/tables-lists-expected.txt

item 1
list item 1
table item 1	table item 2	table item 3
landmark 1
table item 1
This test makes sure that items in lists, tables and landmarks have different traits on them, because they're contained in different items.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS listItem.isInList is true
PASS tableItem.isInTable is true
PASS landmarkItem.isInLandmark is true
PASS nonTableItem.isInTable is false
PASS successfullyParsed is true

TEST COMPLETE
